+++ Regions.py
@@ -1,14 +1,13 @@
 import typing
 
+from worlds.metroidprime.Logic import can_ice_beam, can_missile, can_plasma_beam, can_power_beam, can_super_missile, can_thermal, can_wave_beam, can_xray, has_energy_tanks, has_required_artifact_count
 from worlds.metroidprime.data.ChozoRuins import ChozoRuinsAreaData
 from worlds.metroidprime.data.MagmoorCaverns import MagmoorCavernsAreaData
+from worlds.metroidprime.data.PhazonMines import PhazonMinesAreaData
 from worlds.metroidprime.data.PhendranaDrifts import PhendranaDriftsAreaData
 from worlds.metroidprime.data.RoomNames import RoomName
 from worlds.metroidprime.data.TallonOverworld import TallonOverworldAreaData
-from .Logic import MetroidPrimeLogic as logic
-from BaseClasses import LocationProgressType, Region
-from .Locations import tallon_location_table, magmoor_location_table, mines_location_table, chozo_location_table, \
-    phen_location_table, MetroidPrimeLocation
+from BaseClasses import Region
 if typing.TYPE_CHECKING:
     from . import MetroidPrimeWorld
 
@@ -22,10 +21,7 @@ def create_regions(world: 'MetroidPrimeWorld', final_boss_selection):
     ChozoRuinsAreaData().create_world_region(world)
     MagmoorCavernsAreaData().create_world_region(world)
     PhendranaDriftsAreaData().create_world_region(world)
-
-    phazon_mines = Region("Phazon Mines", world.player, world.multiworld)
-    phazon_mines.add_locations(mines_location_table, MetroidPrimeLocation)
-    world.multiworld.regions.append(phazon_mines)
+    PhazonMinesAreaData().create_world_region(world)
 
     impact_crater = Region("Impact Crater", world.player, world.multiworld)
     world.multiworld.regions.append(impact_crater)
@@ -46,13 +42,16 @@ def create_regions(world: 'MetroidPrimeWorld', final_boss_selection):
     chozo_transport_to_tallon_north = world.multiworld.get_region(RoomName.Transport_to_Tallon_Overworld_North.value, world.player)
     chozo_transport_to_magmoor_north = world.multiworld.get_region(RoomName.Transport_to_Magmoor_Caverns_North.value, world.player)
     chozo_transport_to_tallon_east = world.multiworld.get_region(RoomName.Transport_to_Tallon_Overworld_East.value, world.player)
-    chozo_transport_to_tallon_south = world.multiworld.get_region(RoomName.Transport_to_Tallon_Overworld_South.value, world.player)
+    chozo_transport_to_tallon_south = world.multiworld.get_region("Chozo Ruins: " + RoomName.Transport_to_Tallon_Overworld_South.value, world.player)
 
     magmoor_transport_to_chozo_north = world.multiworld.get_region(RoomName.Transport_to_Chozo_Ruins_North.value, world.player)
     magmoor_transport_to_phazon_west = world.multiworld.get_region(RoomName.Transport_to_Phazon_Mines_West.value, world.player)
     magmoor_transport_to_phendrana_north = world.multiworld.get_region(RoomName.Transport_to_Phendrana_Drifts_North.value, world.player)
     magmoor_transport_to_phendrana_south = world.multiworld.get_region(RoomName.Transport_to_Phendrana_Drifts_South.value, world.player)
     magmoor_transport_to_tallon_west = world.multiworld.get_region(RoomName.Transport_to_Tallon_Overworld_West.value, world.player)
+
+    phazon_mines_transport_to_magmoor_south = world.multiworld.get_region("Phazon Mines: " + RoomName.Transport_to_Magmoor_Caverns_South.value, world.player)
+    phazon_mines_transport_to_tallon_south = world.multiworld.get_region("Phazon Mines: " + RoomName.Transport_to_Tallon_Overworld_South.value, world.player)
 
     phendrana_transport_to_magmoor_west = world.multiworld.get_region(RoomName.Transport_to_Magmoor_Caverns_West.value, world.player)
     phendrana_transport_to_magmoor_south = world.multiworld.get_region("Phendrana Drifts: " + RoomName.Transport_to_Magmoor_Caverns_South.value, world.player)  # There are two transports to magmoor south, other is in phazon mines
@@ -61,14 +60,14 @@ def create_regions(world: 'MetroidPrimeWorld', final_boss_selection):
     tallon_transport_to_chozo_east.connect(chozo_transport_to_tallon_east, "East Chozo Elevator")
     tallon_transport_to_chozo_south.connect(chozo_transport_to_tallon_south, "South Chozo Elevator")
     tallon_transport_to_magmoor_east.connect(magmoor_transport_to_tallon_west, "East Magmoor Elevator")
-    tallon_transport_to_phazon_east.connect(phazon_mines, "East Mines Elevator")
+    tallon_transport_to_phazon_east.connect(phazon_mines_transport_to_tallon_south, "East Mines Elevator")
 
     chozo_transport_to_tallon_north.connect(tallon_transport_to_chozo_west, "North Tallon Elevator")
     chozo_transport_to_tallon_east.connect(tallon_transport_to_chozo_east, "East Tallon Elevator")
     chozo_transport_to_tallon_south.connect(tallon_transport_to_chozo_south, "South Tallon Elevator")
 
     magmoor_transport_to_chozo_north.connect(chozo_transport_to_magmoor_north, "North Chozo Elevator")
-    magmoor_transport_to_phazon_west.connect(phazon_mines, "West Mines Elevator")
+    magmoor_transport_to_phazon_west.connect(phazon_mines_transport_to_magmoor_south, "West Mines Elevator")
     magmoor_transport_to_phendrana_north.connect(phendrana_transport_to_magmoor_west, "North Phendrana Elevator")
     magmoor_transport_to_phendrana_south.connect(phendrana_transport_to_magmoor_south, "South Phendrana Elevator")
     magmoor_transport_to_tallon_west.connect(tallon_transport_to_magmoor_east, "West Tallon Elevator")
@@ -76,24 +75,24 @@ def create_regions(world: 'MetroidPrimeWorld', final_boss_selection):
     phendrana_transport_to_magmoor_west.connect(magmoor_transport_to_phendrana_north, "West Magmoor Elevator")
     phendrana_transport_to_magmoor_south.connect(magmoor_transport_to_phendrana_south, "South Magmoor Elevator")
 
+    artifact_temple = world.multiworld.get_region(RoomName.Artifact_Temple.value, world.player)
+
     if final_boss_selection == 0 or final_boss_selection == 2:
         world.multiworld.get_region(RoomName.Artifact_Temple.value, world.player).connect(impact_crater, "Crater Access", lambda state: (
-            logic.prime_has_missiles(state, world.multiworld, world.player) and
-            logic.prime_artifact_count(state, world.multiworld, world.player) and
-            state.count('Energy Tank', world.player) >= 8 and
-            state.has_all({"Wave Beam", "Ice Beam", "Plasma Beam", "Thermal Visor", "X-Ray Visor", "Phazon Suit",
-                           "Space Jump Boots"}, world.player)))
+            can_missile(state, world.player) and
+            has_required_artifact_count(state, world.player) and
+            has_energy_tanks(state, world.player, 8) and
+            can_plasma_beam(state, world.player) and can_wave_beam(state, world.player) and can_ice_beam(state, world.player) and can_power_beam(state, world.player) and
+            can_xray(state, world.player, True) and can_thermal(state, world.player, True)))
     elif final_boss_selection == 1:
-        world.multiworld.get_region(RoomName.Artifact_Temple.value, world.player).connect(mission_complete, "Mission Complete", lambda state: (
-            logic.prime_has_missiles(state, world.multiworld, world.player) and
-            logic.prime_artifact_count(state, world.multiworld, world.player) and
-            (state.has("Plasma Beam", world.player) or logic.prime_can_super(state, world.multiworld,
-                                                                             world.player)) and
-            logic.prime_etank_count(state, world.multiworld, world.player) >= 8))
+        world.multiworld.get_region(RoomName.Artifact_Temple.value, world.player).connect(mission_complete, "Mission Complete", lambda state:
+                                                                                          can_missile(state, world.player) and
+                                                                                          has_required_artifact_count(state, world.player) and (can_plasma_beam(state, world.player) or can_super_missile(state, world.multiworld, world.player)) and
+                                                                                          has_energy_tanks(state, world.player, 8))
     elif final_boss_selection == 3:
         world.multiworld.get_region(RoomName.Artifact_Temple.value, world.player).connect(mission_complete, "Mission Complete", lambda state: (
-            logic.prime_has_missiles(state, world.multiworld, world.player) and
-            logic.prime_artifact_count(state, world.multiworld, world.player)))
+            can_missile(state, world.player) and
+            has_required_artifact_count(state, world.player)))
 
     if (final_boss_selection == 0 or
             final_boss_selection == 2):
