+++ data/TallonOverworld.py
@@ -1,5 +1,5 @@
 
-from worlds.metroidprime.Logic2 import can_bomb, can_boost, can_charge_beam, can_crashed_frigate, can_crashed_frigate_backwards, can_grapple, can_missile, can_morph_ball, can_move_underwater, can_power_bomb, can_space_jump, can_spider, can_super_missile, can_xray
+from worlds.metroidprime.Logic import can_bomb, can_boost, can_charge_beam, can_crashed_frigate, can_crashed_frigate_backwards, can_grapple, can_missile, can_morph_ball, can_move_underwater, can_power_bomb, can_space_jump, can_spider, can_super_missile, can_xray
 from worlds.metroidprime.data.Tricks import Tricks
 from .RoomData import AreaData, DoorData, DoorLockType, MetroidPrimeArea, PickupData, RoomData
 from .RoomNames import RoomName
